+++ BackEnd/Matmassig/inventory-service/src/main/java/com/school/matmassig/inventoryservice/repository/InventoryItemRepository.java
@@ -2,9 +2,8 @@ package com.school.matmassig.inventoryservice.repository;
 
 import com.school.matmassig.inventoryservice.model.InventoryItem;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
-public interface InventoryItemRepository extends JpaRepository<InventoryItem, Long> {
-    List<InventoryItem> findByUserId(Long userId);
+public interface InventoryItemRepository extends JpaRepository<InventoryItem, Integer> {
+    List<InventoryItem> findByUserId(Integer userId);
 }
